backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mangum==0.17.0
asyncpg==0.29.0
sqlalchemy==2.0.23
supabase==2.0.2
pyjwt==2.8.0
stripe==7.3.0
python-digitalocean==1.17.0
boto3==1.34.0
httpx>=0.24.0,<0.25.0
pydantic==2.5.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
structlog==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
